+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,12 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
-import kitchenpos.dao.MenuDao;
-import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.OrderLineItemDao;
+import kitchenpos.menu.domain.dao.MenuDao;
+import kitchenpos.order.domain.dao.OrderDao;
+import kitchenpos.order.domain.dao.OrderLineItemDao;
 import kitchenpos.table.domain.dao.OrderTableDao;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
