+++ src/main/java/kitchenpos/order/domain/dao/JdbcTemplateOrderDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain.dao;
 
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
