+++ src/main/java/kitchenpos/order/domain/dao/OrderDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain.dao;
 
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 
 import java.util.List;
 import java.util.Optional;
