+++ src/main/java/kitchenpos/order/presentation/OrderRestController.java
@@ -1,12 +1,16 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.OrderService;
-import kitchenpos.domain.Order;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package kitchenpos.order.presentation;
 
 import java.net.URI;
 import java.util.List;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.application.OrderService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class OrderRestController {
