+++ src/test/java/kitchenpos/dao/JdbcTemplateMenuGroupDaoTest.java
@@ -4,7 +4,7 @@ import static kitchenpos.support.fixture.domain.MenuGroupFixture.KOREAN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
