+++ src/test/java/kitchenpos/dao/JdbcTemplateMenuProductDaoTest.java
@@ -8,9 +8,9 @@ import static kitchenpos.support.fixture.domain.ProductFixture.APPLE_1000;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
 import kitchenpos.support.fixture.domain.ProductFixture;
 import org.junit.jupiter.api.BeforeEach;
