+++ src/test/java/kitchenpos/dao/JdbcTemplateOrderDaoTest.java
@@ -5,8 +5,8 @@ import static kitchenpos.support.fixture.domain.TableGroupFixture.getTableGroup;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.support.fixture.domain.OrderFixture;
