+++ src/test/java/kitchenpos/dao/JdbcTemplateOrderLineItemDaoTest.java
@@ -9,10 +9,10 @@ import static kitchenpos.support.fixture.domain.OrderTableFixture.GUEST_ONE_EMPT
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.support.fixture.domain.TableGroupFixture;
