+++ src/test/java/kitchenpos/dao/JdbcTemplateTest.java
@@ -1,6 +1,11 @@
 package kitchenpos.dao;
 
 import javax.sql.DataSource;
+import kitchenpos.menu.domain.dao.JdbcTemplateMenuDao;
+import kitchenpos.menu.domain.dao.JdbcTemplateMenuGroupDao;
+import kitchenpos.menu.domain.dao.JdbcTemplateMenuProductDao;
+import kitchenpos.order.domain.dao.JdbcTemplateOrderDao;
+import kitchenpos.order.domain.dao.JdbcTemplateOrderLineItemDao;
 import kitchenpos.product.domain.dao.JdbcTemplateProductDao;
 import kitchenpos.support.tool.DataBaseCleaner;
 import kitchenpos.table.domain.dao.JdbcTemplateOrderTableDao;
