+++ src/test/java/kitchenpos/support/fixture/domain/MenuFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.support.fixture.domain;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 public enum MenuFixture {
