+++ src/test/java/kitchenpos/support/fixture/domain/MenuProductFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.support.fixture.domain;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public enum MenuProductFixture {
